requirements.txt
torch
torchvision
matplotlib
Pillow
pandas
pyarrow
tqdm
